Created dataload utility module
