+++ applications/harnesses/adams_lsc_policy/train_lsc_policy.py
@@ -11,6 +11,7 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 from yoke.models.policyCNNmodules import gaussian_policyCNN
 from yoke.datasets.lsc_dataset import LSC_hfield_policy_DataSet
 import yoke.torch_training_utils as tr
+from yoke.utils.dataload import make_distributed_dataloader
 from yoke.utils.checkpointing import save_model_and_optimizer
 from yoke.utils.checkpointing import load_model_and_optimizer
 from yoke.lr_schedulers import CosineWithWarmupScheduler
@@ -240,7 +241,7 @@ def main(
     )
 
     # NOTE: For DDP the batch_size is the per-GPU batch_size!!!
-    train_dataloader = tr.make_distributed_dataloader(
+    train_dataloader = make_distributed_dataloader(
         train_dataset,
         batch_size,
         shuffle=True,
@@ -248,7 +249,7 @@ def main(
         rank=rank,
         world_size=world_size,
     )
-    val_dataloader = tr.make_distributed_dataloader(
+    val_dataloader = make_distributed_dataloader(
         val_dataset,
         batch_size,
         shuffle=False,
