+++ applications/harnesses/ch_lsc_reward/train_lsc_reward.py
@@ -21,7 +21,8 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 
 from yoke.models.hybridCNNmodules import hybrid2vectorCNN
 from yoke.datasets.lsc_dataset import LSC_hfield_reward_DataSet
-import yoke.torch_training_utils as tr  # Assuming this exists
+import yoke.torch_training_utils as tr
+from yoke.utils.dataload import make_distributed_dataloader
 from yoke.utils.checkpointing import load_model_and_optimizer
 from yoke.utils.checkpointing import save_model_and_optimizer
 from yoke.lr_schedulers import CosineWithWarmupScheduler
@@ -31,7 +32,8 @@ from yoke.helpers import cli  # Assuming this exists
 # Inputs
 #############################################
 descr_str = (
-    "Trains reward network architecture to calculate error between current and target density fields."
+    "Trains reward network architecture to calculate error between current and target "
+    "density fields."
 )
 parser = argparse.ArgumentParser(
     prog="reward network training", description=descr_str, fromfile_prefix_chars="@"
@@ -258,7 +260,7 @@ def main(
     print("Datasets initialized.")
 
     # NOTE: For DDP the batch_size is the per-GPU batch_size!!!
-    train_dataloader = tr.make_distributed_dataloader(
+    train_dataloader = make_distributed_dataloader(
         train_dataset,
         batch_size,
         shuffle=True,
@@ -266,7 +268,7 @@ def main(
         rank=rank,
         world_size=world_size,
     )
-    val_dataloader = tr.make_distributed_dataloader(
+    val_dataloader = make_distributed_dataloader(
         val_dataset,
         batch_size,
         shuffle=False,
